+++ src/business/stego/PNGStego.h
@@ -3,14 +3,14 @@
 #include "LSBStegoBusiness.h"
 class PNGStego:public LSBStegoBusiness
 {
-private:
-    unsigned int getTransparentPixels();
+private:	
+    unsigned long int getTransparentPixels();
 public:
 	PNGStego();
 	PNGStego(std::string filename);
-	virtual unsigned int changePixel(BYTE *pixels,const char* mensaje, long int size);
-	virtual std::string getMessageFromPixel(BYTE *pixels,unsigned int longitud);
-	virtual unsigned int getFreeSpace();
+	virtual unsigned long int changePixel(BYTE *pixels,const char* mensaje,unsigned long int size);
+	virtual std::string getMessageFromPixel(BYTE *pixels,unsigned long int longitud);
+	virtual unsigned long int getFreeSpace();
 	virtual ~PNGStego();
 };
